refactor(Team): Team 클래스를 stone 패키지에서 밖으로 꺼냄
