+++ domain/src/main/kotlin/domain/state/GameState.kt
@@ -3,7 +3,7 @@ package domain.state
 import domain.Board
 import domain.stone.Point
 import domain.stone.Stone
-import domain.stone.Team
+import domain.Team
 
 abstract class GameState {
     abstract val turn: Team
